Copying //UE4/Release-4.25Plus-Dev to Release-4.25Plus (//UE4/Release-4.25Plus) #rb none #jira none
#ROBOMERGE-OWNER: jeff.newquist
#ROBOMERGE-AUTHOR: jeff.newquist
#ROBOMERGE-SOURCE: CL 13740250 in //UE4/Release-4.25Plus/...
#ROBOMERGE-BOT: RELEASE (Release-4.25Plus -> Release-Engine-Staging) (v708-13641880)

[CL 13742073 by jeff newquist in Release-Engine-Staging branch]

+++ Engine/Source/Runtime/NetworkFile/Private/NetworkPlatformFile.cpp
@@ -954,19 +954,22 @@ public:
 	/** Write the file  */
 	void DoWork()
 	{
-		if (InnerPlatformFile.FileExists(*Filename))
-		{
-			InnerPlatformFile.SetReadOnly(*Filename, false);
-			InnerPlatformFile.DeleteFile(*Filename);
-		}
 		// Read FileSize first so that the correct amount of data is read from the archive
 		// before exiting this worker.
 		uint64 FileSize;
 		*FileArchive << FileSize;
+
+		bool bCopiedExternally = (FileSize == MAX_uint64); // -1 filesize means that we already copied it via TargetPlatform
+
+		if (!bCopiedExternally && InnerPlatformFile.FileExists(*Filename) ) 
+		{
+			InnerPlatformFile.SetReadOnly(*Filename, false);
+			InnerPlatformFile.DeleteFile(*Filename);
+		}
+
 		if (ServerTimeStamp != FDateTime::MinValue())  // if the file didn't actually exist on the server, don't create a zero byte file
 		{
-			// -1 filesize means that we already copied it via TargetPlatform
-			if (FileSize != MAX_uint64)
+			if (!bCopiedExternally)
 			{
 				FString TempFilename = Filename + TEXT(".tmp");
 				InnerPlatformFile.CreateDirectoryTree(*FPaths::GetPath(Filename));
